1 bug fix 3 enhancements 1 documentation issue. See summary for full feature details
- Issue #18 Bug: Fixed, separated the point size and jitter width, these
are not independent. Jitter width only appears when jitter is selected.
- Issue #11 Enhancement: This was a 2 part fix, first we fixed the issue
with allowing users to set a factor as Y and a numeric as X and just
behind the scenes switching it for them. Instead we now switch the X and
Y for them in the select boxes. This allowed us to simply add a
coord_flip to the ggplot object.
Issue #9: Enhancement: Another big one. Only Survey 2 Respondents has
been added to the exploratory bar/point/violin/box tab with some generic
code that will disable (per request instead of hiding it) the checkbox
if it wouldn't result in a change.
Issue #8 Documentation added
Issue #6 - fixed