Working Shiny Authentication
NOTE: A hash code of the password is required to make this app run in
its current state.. It lives in global.R. Be smart, even at SHA512 you
shouldn't commit your hash codes to git or any other repository. ;)
NOTE: There may be ways to bypass this security through some sort of
injection or java script variable altering. This is just the lock that
keeps honest people honest. None of the graphs or variable drop downs
should function without the authentication which changes server side
variables, but there are plenty of people that know more about shiny and
R security than I do. Any comments on better ways to add authentication
through the shiny app itself are welcome.
Note: Only run this on HTTPS connections, otherwise it's possible Shiny
will transmit the pin in the clear since the hashing takes place on the
server.